Version 2.0!
This version accomplishes the same end-goal with a more intelligent design.  It is no longer necessary to create customized macros.

Explanation:

The client will not allow a cast while the cast bar is being displayed.  Ordinarily, the cast bar will continue to be displayed until the server informs the client that the cast has completed.  Version 2.0 alters that behavior, causing the cast bar to disappear automatically once a sufficient amount of time has passed.  Though this solution requires a greater number of modifications to the game, it essentially eliminates potential for abuse or misuse, which should make it less ethically ambiguous by those who are concerned.